requirements.txt
Django==4.2.7
mysqlclient==2.2.0
python-decouple==3.8
Pillow==10.1.0
django-crispy-forms==2.1
crispy-bootstrap4==2022.1